model/model_serving/requirements.txt
functions-framework==3.2.0
google-cloud-storage==2.5.0
matplotlib==3.6.1
mlflow==1.30.0
numpy==1.23.1
pandas==1.4.3
requests==2.25.0
scikit-learn==1.1.2
scipy==1.9.0
